+++ packages/terra-date-picker/src/terra-dev-site/test/date-picker/DatePickerMinMax.test.jsx
@@ -4,8 +4,8 @@ import DatePicker from '../../../../lib/DatePicker';
 const DatePickerMinMax = () => (
   <DatePicker
     name="date-input"
-    minDate={'2017-04-01'}
-    maxDate={'2017-04-10'}
+    minDate="2017-04-01"
+    maxDate="2017-04-10"
   />
 );