+++ packages/terra-dialog/src/terra-dev-site/doc/example/DialogDefault.jsx
@@ -7,10 +7,12 @@ const header = 'Header Stuff';
 
 const footer = <span>Footer Stuff</span>;
 
-const body = (<div>
-  <p>This is my body content.</p>
-  <p>This is some more content.</p>
-</div>);
+const body = (
+  <div>
+    <p>This is my body content.</p>
+    <p>This is some more content.</p>
+  </div>
+);
 
 const DialogDefault = () => (
   <div style={{ height: '200px', width: '350px', border: 'dashed' }}>