+++ packages/terra-form-fieldset/src/terra-dev-site/test/form-fieldset/PopulatedFieldset.test.jsx
@@ -3,7 +3,7 @@ import React from 'react';
 import Fieldset from '../../../Fieldset';
 
 const fieldset = () =>
-  <Fieldset
+  (<Fieldset
     id="populated-fieldset"
     legend="Do you have any children?"
     legendAttrs={{ className: 'healtheintent-application' }}
@@ -11,6 +11,6 @@ const fieldset = () =>
     value="children_present"
     help="Families are eligible for family package plans"
     required
-  />;
+  />);
 
 export default fieldset;